tutorials/gemini_embeddings/requirements.txt
google-genai>=1.0.0
chromadb>=0.6.0
numpy>=1.24.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
rich>=13.0.0
gradio>=4.0.0
requests>=2.28.0
